+++ frontend/src/views/organisation/workflows/MoOrganisationUnitMove.vue
@@ -17,6 +17,7 @@
             v-model="original"
             :label="$t('input_fields.select_unit')"
             :date="move.data.validity.from"
+            :validity="validity"
             required
           />
         </div>
@@ -37,6 +38,8 @@
         v-model="move.data.parent"
         :label="$t('input_fields.select_new_super_unit')"
         :date="move.data.validity.from"
+        :validity="validity"
+        :extra-validations="parentValidations"
         required
       />
 
@@ -101,6 +104,21 @@ export default {
       isLoading: false,
       backendValidationError: null
     }
+  },
+
+  computed: {
+    validity () {
+      return {
+        'from': this.move.data.validity.from
+      }
+    },
+
+    parentValidations () {
+      return {
+        candidate_parent_org_unit: [this.original, this.move.data.parent, this.validity]
+      }
+    }
+
   },
 
   watch: {