+++ NEWS.rst
@@ -5,6 +5,8 @@ New features
 ------------
 
 * #27687: The various org funk relations now accept a ``user_key`` parameter.
+* #25396: Implemented validation of individual fields in frontend using
+  backend validation API.
 
 Bug fixes
 ---------